main/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
python-multipart>=0.0.9,<1.0
jinja2>=3.1,<4.0
httpx>=0.27,<1.0
openai>=1.0
redis>=5.0.0
SQLAlchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
pydantic-settings>=2.6,<3.0
sentence-transformers==5.3.0
transformers>=4.57.0,<5.0
qwen-vl-utils>=0.0.14,<1.0
